QA 2.0.0.2 to main

+++ src/app/modules/project-hub/common/key-assumptions-table/key-assumptions-table.component.ts
@@ -1,8 +1,9 @@
 import {Component, Input, OnInit} from '@angular/core';
 import {ProjectApiService} from "../project-api.service";
-import {ActivatedRoute} from "@angular/router";
+import {ActivatedRoute, Router} from "@angular/router";
 import {ProjectHubService} from "../../project-hub.service";
 import {FuseConfirmationConfig, FuseConfirmationService} from "../../../../../@fuse/services/confirmation";
+import {Constants} from "../../../../shared/constants";
 
 @Component({
     selector: 'app-key-assumptions-table',
@@ -10,13 +11,15 @@ import {FuseConfirmationConfig, FuseConfirmationService} from "../../../../../@f
     styleUrls: ['./key-assumptions-table.component.scss']
 })
 export class KeyAssumptionsTableComponent implements OnInit {
-    @Input() callLocation:  'Normal'  | 'Project-Charter'  = 'Normal'
+    @Input() callLocation:  'Normal'  | 'Project-Charter' | 'Business-Case'  = 'Normal'
     @Input() editable: boolean
     keyAssumptions: any = []
     id: string = ''
     isGrid: boolean = false
+    keyAssumptionsViewEditType: string = "KeyAssumptions";
+    keyAssumptionsBulkEditType: string = "KeyAssumptionsBulkEdit";
     constructor(private apiService: ProjectApiService, private _Activatedroute: ActivatedRoute, public projecthubservice: ProjectHubService
-    ,public fuseAlert: FuseConfirmationService) {
+    ,public fuseAlert: FuseConfirmationService, private router: Router) {
         this.projecthubservice.submitbutton.subscribe(res => {
             if (res == true) {
                 this.dataloader()
@@ -29,12 +32,34 @@ export class KeyAssumptionsTableComponent implements OnInit {
     }
     dataloader(){
         this.id = this._Activatedroute.parent.snapshot.paramMap.get("id");
-        if(this.callLocation != 'Normal'){
+        if(this.callLocation == 'Project-Charter'){
             this.id = this._Activatedroute.parent.parent.snapshot.paramMap.get("id");
+            this.keyAssumptionsViewEditType = "ProjectCharterKeyAssumptionAddSingle"
+            this.keyAssumptionsBulkEditType = "ProjectCharterKeyAssumptionBulkEdit"
+            this.apiService.getKeyAssumptionsByProject(this.id).then((res) => {
+                this.keyAssumptions = res
+            })
         }
-        this.apiService.getKeyAssumptionsByProject(this.id).then((res) => {
-            this.keyAssumptions = res
-        })
+        if(this.callLocation == 'Business-Case'){
+            this.id = this._Activatedroute.parent.parent.parent.snapshot.paramMap.get("id")
+            this.keyAssumptionsViewEditType = "BusinessCaseKeyAssumptionAddSingle"
+            this.keyAssumptionsBulkEditType = "BusinessCaseKeyAssumptionBulkEdit"
+            if (this.router.url.includes('recommended-option')) {
+                this.apiService.getKeyAssumptionsByProject(this.id).then((res) => {
+                    this.keyAssumptions = res
+                })
+            }
+            if (this.router.url.includes('option-2')) {
+                this.apiService.getKeyAssumptionsByOption(this.id,Constants.OPTION_2_ID.toString()).then((res) => {
+                    this.keyAssumptions = res
+                })
+            }
+            if (this.router.url.includes('option-3')) {
+                this.apiService.getKeyAssumptionsByOption(this.id,Constants.OPTION_3_ID.toString()).then((res) => {
+                    this.keyAssumptions = res
+                })
+            }
+        }
     }
     deleteKeyAssumption(id: string) {
         var comfirmConfig: FuseConfirmationConfig = {
@@ -59,12 +84,25 @@ export class KeyAssumptionsTableComponent implements OnInit {
             "dismissible": true
         }
         const keyAsumptioneAlert = this.fuseAlert.open(comfirmConfig)
-
         keyAsumptioneAlert.afterClosed().subscribe(close => {
             if (close == 'confirmed') {
-                this.apiService.deleteKeyAssumption(id).then(res => {
-                    this.projecthubservice.submitbutton.next(true)
-                })
+                if (this.callLocation == 'Business-Case') {
+                    if (this.router.url.includes('recommended-option')) {
+                        this.apiService.deleteKeyAssumption(id).then(res => {
+                            this.projecthubservice.submitbutton.next(true)
+                        })
+                    }
+                    if (this.router.url.includes('option-2') || this.router.url.includes('option-3')) {
+                        this.apiService.deleteKeyAssumptionByOption(id).then((res) => {
+                            this.projecthubservice.submitbutton.next(true)
+                        })
+                    }
+                }
+                if (this.callLocation == "Project-Charter") {
+                    this.apiService.deleteKeyAssumption(id).then(res => {
+                         this.projecthubservice.submitbutton.next(true)
+                    })
+                }
             }
         })
     }
